requirements.txt
anyio==3.6.2
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
astroid==2.15.4
autopep8==2.0.2
boto3==1.26.126
botocore==1.29.126
certifi==2018.1.18
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
cssselect2==0.7.0
defusedxml==0.7.1
dill==0.3.6
distlib==0.3.6
dj-database-url==0.4.2
Django==4.2
django-allauth==0.54.0
django-autoslug==1.9.9
django-cleanup==7.0.0
django-cors-headers==3.14.0
django-countries==7.5.1
django-crispy-forms==2.0
django-filter==23.2
django-material-admin==1.8.6
django-phonenumber-field==7.1.0
django-pwa==1.1.0
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
fastapi==0.95.0
filelock==3.9.0
freetype-py==2.3.0
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
html5lib==1.1
idna==2.6
isort==5.12.0
jmespath==1.0.1
lazy-object-proxy==1.9.0
lxml==4.9.2
material==0.1
mccabe==0.7.0
mysqlclient==2.1.1
oauthlib==3.2.2
oscrypto==1.3.0
Pillow==9.4.0
platformdirs==3.0.0
psycopg2==2.9.5
pycairo==1.23.0
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.7
pyHanko==0.18.1
pyhanko-certvalidator==0.22.0
PyJWT==2.6.0
pylint==2.17.3
pypdf==3.8.1
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2017.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.4.2
reportlab==4.0.0
requests==2.29.0
requests-oauthlib==1.3.1
rest-framework-simplejwt==0.0.2
rlPyCairo==0.2.0
s3transfer==0.6.0
schedule==0.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.9
sqlparse==0.4.4
starlette==0.26.1
svglib==1.5.1
tinycss2==1.2.1
tomlkit==0.11.8
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
uritools==4.0.1
urllib3==1.26.15
uvicorn==0.21.1
virtualenv==20.19.0
webencodings==0.5.1
whitenoise==6.4.0
wrapt==1.15.0
xhtml2pdf==0.2.10
